adding subreddit-specific testing to generation tests

+++ scripts/data_processing/test_question_generation.py
@@ -45,36 +45,63 @@ def main():
     num_beams = 8
     pred_data = generate_predictions(generation_model, test_data, model_tokenizer,
                                          generation_method=generation_method,
-                                         num_beams=num_beams,
-                                     )
+                                         num_beams=num_beams,)
 
     ## get aggregate scores
+    generated_text_score_out_file = os.path.join(out_dir,
+                                                 'test_data_output_scores.tsv')
+    if(not os.path.exists(generated_text_score_out_file)):
+        generation_score_data = test_question_overlap(pred_data, test_data)
+        # compute mean/sd
+        generation_score_means = generation_score_data.mean(axis=0)
+        generation_score_sd = generation_score_data.std(axis=0)
+        generation_score_data = pd.concat([
+            generation_score_means,
+            generation_score_sd,
+        ], axis=1).transpose()
+        generation_score_data.index = ['mean', 'sd']
+        generation_score_data = generation_score_data.reset_index().rename(columns={'index' : 'stat'})
+
+        ## write things to file
+        generated_text_out_file = os.path.join(out_dir, 'test_data_output_text.gz')
+        with gzip.open(generated_text_out_file, 'wt') as generated_text_out:
+            generated_text_out.write('\n'.join(pred_data))
+        generation_score_data.to_csv(generated_text_score_out_file, sep='\t', index=False)
+
+    ## optional: same thing but for different subsets of post data
+        # if (args.get('post_metadata') is not None):
+        #     post_metadata = pd.read_csv(args['post_metadata'], sep='\t',
+        #                                 compression='gzip', index_col=False,
+        #                                 usecols=['id', 'subreddit'])
+        #     community_var = 'subreddit'
+        #     post_article_ids = np.array(post_metadata['article_id'])
+        #     for community_i, metadata_i in post_metadata.groupby(community_var):
+        #         article_ids_i = set(metadata_i.loc[:, 'article_id'].unique())
+        #         idx_i = np.where(
+        #             np.apply_along_axis(lambda x: x in article_ids_i, 0,
+        #                                 post_article_ids.reshape(1, -1)))
+        #         # idx_i = np.where(post_metadata['article_id'].isin(metadata_i.loc[:, 'id'].unique()))
+        #         test_data_i = test_data.select(idx_i)
+        #         data_name_i = f'{community_i}_{data_name}'
+        #         evaluate_model(model, train_data, test_data_i, tokenizer, out_dir,
+        #                        data_name=data_name_i)
+
+
+def test_question_overlap(pred_data, test_data):
     ## TODO: need semantic overlap score! for input and for target text
     generation_scores = []
-    bleu_weights = [1.0, 0., 0., 0.] # 100% 1-grams, 0% 2-grams, etc.
+    bleu_weights = [1.0, 0., 0., 0.]  # 100% 1-grams, 0% 2-grams, etc.
     rouge_scorer = RougeScorer(['rougeL'], use_stemmer=True)
     for test_data_i, pred_data_i in zip(test_data['target_text'], pred_data):
-        bleu_score_i = compute_text_bleu(test_data_i, pred_data_i, weights=bleu_weights)
+        bleu_score_i = compute_text_bleu(test_data_i, pred_data_i,
+                                         weights=bleu_weights)
         rouge_score_data_i = rouge_scorer.score(test_data_i, pred_data_i)
         rouge_score_i = rouge_score_data_i['rougeL'].fmeasure
         generation_scores.append([bleu_score_i, rouge_score_i])
-    generation_score_data = pd.DataFrame(generation_scores, columns=['BLEU-1', 'ROUGE-L'])
-    # compute mean/sd
-    generation_score_means = generation_score_data.mean(axis=0)
-    generation_score_sd = generation_score_data.std(axis=0)
-    generation_score_data = pd.concat([
-        generation_score_means,
-        generation_score_sd,
-    ], axis=1).transpose()
-    generation_score_data.index = ['mean', 'sd']
-    generation_score_data = generation_score_data.reset_index().rename(columns={'index' : 'stat'})
+    generation_score_data = pd.DataFrame(generation_scores,
+                                         columns=['BLEU-1', 'ROUGE-L'])
+    return generation_score_data
 
-    ## write things to file
-    generated_text_out_file = os.path.join(out_dir, 'test_data_output_text.gz')
-    generated_text_score_out_file = os.path.join(out_dir, 'test_data_output_scores.tsv')
-    with gzip.open(generated_text_out_file, 'wt') as generated_text_out:
-        generated_text_out.write('\n'.join(pred_data))
-    generation_score_data.to_csv(generated_text_score_out_file, sep='\t', index=False)
 
 if __name__ == '__main__':
     main()
